# Módulo 5 - Node.js com express (Avançado)/Node.js com Express (Avançado) - Aula 7 - 19 de fev/aula-07-main (1)/aula-07-main/src/controllers/user.controller.ts
import { Request, Response } from 'express';
import { UserServiceInterface } from '../services/user-repository.interface';


export class UserController {
  private userService: UserServiceInterface
  constructor(userService: UserServiceInterface) {
    this.userService = userService
  }

  async getAllUsers(req: Request, res: Response): Promise<void> {
    try {
      const users = await this.userService.getAllUsers();
      res.json(users);
    } catch (error) {
      console.error(error);
      res.status(500).json({ error: 'Internal Server Error' });
    }
  }

  async getUserById(req: Request, res: Response): Promise<void> {
    const userId = Number(req.params.id);
    try {
      const user = await this.userService.getUserById(userId);
      if (user) {
        res.json(user);
      } else {
        res.status(404).json({ error: 'User not found' });
      }
    } catch (error) {
      console.error(error);
      res.status(500).json({ error: 'Internal Server Error' });
    }
  }

  async createUser(req: Request, res: Response): Promise<void> {
    const { name, email } = req.body;
    try {
      const newUser = await this.userService.createUser(name, email);
      res.status(201).json(newUser);
    } catch (error) {
      console.error(error);
      res.status(500).json({ error: 'Internal Server Error' });
    }
  }

  async deleteUser(req: Request, res: Response): Promise<void> {
    const userId = Number(req.params.id);
    try {
      await this.userService.deleteUser(userId);
      res.json({ message: 'User deleted successfully' });
    } catch (error) {
      console.error(error);
      res.status(500).json({ error: 'Internal Server Error' });
    }
  }
}